Update Calcular Vlocidad Olimpiada.c

+++ Calcular Vlocidad Olimpiada.c
@@ -1,19 +1,19 @@
 //Calcular velocidad en olimpiada
 #include<stdio.h>
 int main(){
-    float dis, cen, seg, min, TSE, VMS, VKH;
+    int dis, cen, seg, min, TSE, VMS, VKH;
     printf("\nDa el valor de la distancia :");
-    scanf("%f", &dis);
+    scanf("%d", &dis);
     printf("\nDa el valor de las centenas: ");
-    scanf("%f", &cen);
+    scanf("%d", &cen);
     printf("\nDa el valor en segundos: ");
-    scanf("%f", &seg);
+    scanf("%d", &seg);
     printf("\nDan el valor en minutos:");
-    scanf("%f", &min);
+    scanf("%d", &min);
     TSE = (min*60+seg+cen)/100;
     VMS = dis/TSE;
     VKH= VMS*3600/1000;
-    printf("\nLa velocidad es: %f KM/HR", VKH);
+    printf("\nLa velocidad es: %d KM/HR", VKH);
 
     return 0;
 }
